JUnit test for several guest users adapted
