+++ src/net/common/sessionmanager.cpp
@@ -34,8 +34,6 @@
 #include <net/serverexception.h>
 #include <net/socket_msg.h>
 
-#include <core/loghelper.h> // debug - remove me
-
 using namespace std;
 
 #define SERVER_MAX_GUEST_USERS_LOBBY	50		// LG: Maximum number of guests users in lobby allowed
